Add a more generic shell pattern matcher
It tries to find a prefix of "% " or "$ " but will look
for any lines wiht those in the line if the prefix isn't found.

The "try anywhere in the line" is likely to have false positives, but is
required for default bash/zsh prompts which have other text on the same
line as the prompt.